Minor cleanups and formatting

+++ apps/probability/gui/statistic_curve_view.cpp
@@ -1,6 +1,6 @@
 #include "statistic_curve_view.h"
 
-#include <math.h>
+#include <cmath>
 #include <poincare/print_float.h>
 
 #include "probability/app.h"
@@ -66,8 +66,8 @@ void StatisticCurveView::colorUnderCurve(KDContext * ctx, KDRect rect,
     return;
   }
 
-  float min = fminf(z, zAlpha);
-  float max = fmaxf(z, zAlpha);
+  float min = std::fmin(z, zAlpha);
+  float max = std::fmax(z, zAlpha);
   bool zMoreExtreme = op == HypothesisParams::ComparisonOperator::Higher ? z > zAlpha : z < zAlpha;
   KDColor middleColor = !zMoreExtreme ? Escher::Palette::YellowDark : Escher::Palette::GrayMiddle;
   KDColor externColor = zMoreExtreme ? Escher::Palette::YellowDark : Escher::Palette::GrayMiddle;
